fixed TODO

+++ src/DataTypes.java
@@ -1,7 +1,9 @@
 public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(int[] numbers) {
-        int s = 0; // variable to accumulate the sum in!
+        // Java has a 32 byte memory limit for ints so need to change to long
+        // used to be int
+        long s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
